refactor: extract presentation logic and fix upload state reset
- Extract getStatusClass and getMimeTypeIcon from BulkFileCard to lib/bulkUpload
- Move handleSingleFile logic into useFileSelect and remove inline callback from FileSelect
- Fix useMetadataUpload: wrap generateSingleFileMetadata in try/catch/finally
  so isUploading and uploadProgress always reset even on error
