+++ hooks/useFileSelect.ts
@@ -5,54 +5,45 @@ import { handleImageSelection } from "@/lib/fileSelect/handleImageSelection";
 import { handleOtherFileSelection } from "@/lib/fileSelect/handleOtherFileSelection";
 import { useMetadataFormProvider } from "@/providers/MetadataFormProvider";
 
-/**
- * Hook for file selection only - stores files as blobs.
- * Upload logic is handled in useMetadataUpload.generateMetadataUri()
- */
 const useFileSelect = () => {
   const { setMimeType, setImageFile, setPreviewFile, setAnimationFile, imageFile } =
     useMetadataFormProvider();
 
-  const selectFile = useCallback(
-    async (event: any) => {
-      const file: File = event.target.files[0];
-      if (!file || !validateFile(file)) {
-        return;
-      }
+  const handleSingleFile = useCallback(
+    async (file: File) => {
+      try {
+        if (!validateFile(file)) return;
+        const mimeType = file.type;
+        const isImage = mimeType.includes("image");
+        const isVideo = mimeType.includes("video");
 
-      const mimeType = file.type;
-      const isImage = mimeType.includes("image");
-      const isVideo = mimeType.includes("video");
-
-      // Store files as blobs - no upload happens here
-      if (isVideo) {
-        await handleVideoSelection(file, {
-          setAnimationFile,
-          setMimeType,
-          setPreviewFile,
-        });
-      } else if (isImage) {
-        await handleImageSelection(file, {
-          setMimeType,
-          setImageFile,
-          setPreviewFile,
-        });
-      } else {
-        await handleOtherFileSelection(file, {
-          setMimeType,
-          setAnimationFile,
-          setPreviewFile,
-        });
-        if (imageFile) {
-          setPreviewFile(imageFile);
-          setImageFile(null);
+        if (isVideo) {
+          await handleVideoSelection(file, { setAnimationFile, setMimeType, setPreviewFile });
+        } else if (isImage) {
+          await handleImageSelection(file, { setMimeType, setImageFile, setPreviewFile });
+        } else {
+          await handleOtherFileSelection(file, { setMimeType, setAnimationFile, setPreviewFile });
+          if (imageFile) {
+            setPreviewFile(imageFile);
+            setImageFile(null);
+          }
         }
+      } catch {
+        // validateFile already shows toast
       }
     },
     [setMimeType, setImageFile, setPreviewFile, setAnimationFile, imageFile]
   );
 
-  return { selectFile };
+  const selectFile = useCallback(
+    async (event: any) => {
+      const file: File = event.target.files[0];
+      if (file) await handleSingleFile(file);
+    },
+    [handleSingleFile]
+  );
+
+  return { selectFile, handleSingleFile };
 };
 
 export default useFileSelect;
