+++ lib/bulkUpload/getMimeTypeIcon.ts
@@ -1,0 +1,5 @@
+export const getMimeTypeIcon = (mimeType: string): string => {
+  if (mimeType.includes("pdf")) return "PDF";
+  if (mimeType.includes("audio")) return "♪";
+  return "?";
+};
